isSundaySortingAllowed also checks on saturdays
The isSundaySortingAllowed function checks if sunday is selected in the config. But also checks on key '6' and '7', so according to the WeekDays config provider sunday is always '0'. 

Now if clients have saturday sorting selected in backend also sunday is allowed and will be used for ship_at date, which is not correct This will fix it.  

'7' is never used, so removed.