Selected preferred delivery date is overwritten by setDefaultData
The date that is posted in the ajax save in the checkout (postnl/deliveryoptions/save/) is overwritten by the setDefaultData observer. In our case this causes the dates to be set to the first available delivery date instead of the selected delivery date (further into the future) in the checkout. 

No time to test on default Magento and to test other cases. So this is just a quick fix that i want to share.