AddressHelper when street lines are < 2
Hey there,

We've experienced a bug with this version of the PostNL module in combination with Magento 2.3.6-p1 regarding our customers that have configured their checkout to only have one street line as is possible in Stores > Configuration > Customers > Customer Configuration > Name and Address Options > Number of lines in a Street Address. In those cases there is no checkout.steps.shipping-step.shippingAddress.shipping-address-fieldset.street.1 field in the uiRegistry which causes the large uiRegistry.get call to not trigger and therefor the address to not be filled.

We've managed to fix this by adding the housenumber seperatly (as you've already done for some other fields) and then checking for the existence of the house number in the large uiRegistry field. We've also decided to fitler the housenumber from the street because that seemed to cause some issues in some specific cases. This might not be necessary.